Paymetn added

+++ app/src/main/java/in/shpt/adapter/ShoppingCartProductAdapter.kt
@@ -5,17 +5,21 @@ import `in`.shpt.activity.ProductDetail
 import `in`.shpt.event.ItemRemovedFromCartEvent
 import `in`.shpt.ext.getIcon
 import `in`.shpt.ext.loadUrl
+import `in`.shpt.ext.modifyCart
 import `in`.shpt.ext.removeFromCart
 import `in`.shpt.widget.Ripple
 import android.app.Activity
+import android.content.DialogInterface.OnShowListener
 import android.content.Intent
 import android.graphics.Color
 import android.os.AsyncTask
 import android.support.v4.os.AsyncTaskCompat
+import android.support.v7.app.AlertDialog
 import android.support.v7.widget.PopupMenu
 import android.support.v7.widget.RecyclerView
 import android.view.MenuItem
 import android.view.View
+import android.widget.EditText
 import android.widget.ImageButton
 import android.widget.ImageView
 import android.widget.TextView
@@ -24,6 +28,8 @@ import com.mikepenz.fastadapter.items.AbstractItem
 import com.mikepenz.fontawesome_typeface_library.FontAwesome
 import org.greenrobot.eventbus.EventBus
 import org.json.JSONObject
+import java.util.*
+
 
 /**
  * Created by poovarasanv on 15/11/16.
@@ -42,15 +48,98 @@ class ShoppingCartProductAdapter(
 
         when (item!!.itemId) {
             R.id.delete_item -> deleteFromCart()
+            R.id.modify_item -> modifyItem()
             else -> context.toast("clicked")
         }
         // context.toast("Clicked...")
-        return true;
+        return true
+    }
+
+    fun modifyItem() {
+        if (context.isConnected()) {
+            var quantityInp: String = ""
+            var promptsView: View = context.inflate(R.layout.quantity_prompt)
+            val alertDialogBuilder = AlertDialog.Builder(context)
+
+            // set prompts.xml to alertdialog builder
+            alertDialogBuilder.setView(promptsView)
+
+            val userInput = promptsView
+                    .findViewById(R.id.input_quantity) as EditText
+
+            userInput.setText(quantity)
+            userInput.setSelection(userInput.text.length)
+
+            alertDialogBuilder
+                    .setCancelable(true)
+                    .setPositiveButton("OK", null)
+                    .setNegativeButton("Cancel"
+                    ) { dialog, id -> dialog.cancel() }
+
+
+            val alertDialog = alertDialogBuilder.create()
+
+            alertDialog.setOnShowListener(OnShowListener {
+                val b = alertDialog.getButton(AlertDialog.BUTTON_POSITIVE)
+                b.setOnClickListener {
+                    quantityInp = userInput.text.toString()
+                    if (quantityInp.isEmpty() || quantityInp.isBlank()) {
+                        userInput.error = "Invalid Quantity"
+                    } else {
+                        modifyCart(quantityInp)
+                        alertDialog.dismiss()
+                    }
+                }
+            })
+            alertDialog.show()
+
+        } else {
+            context.toast("No INternet Connection")
+        }
+    }
+
+    fun modifyCart(quantity: String) {
+        var map: HashMap<String, String> = HashMap()
+        var qu: String = "quantity[$productId::]"
+        map.put(qu, quantity)
+
+        AsyncTaskCompat.executeParallel(ModifyCart(), map)
+    }
+
+    inner class ModifyCart : AsyncTask<HashMap<String, String>, Void, JSONObject>() {
+        override fun doInBackground(vararg p0: HashMap<String, String>?): JSONObject? {
+            return context.modifyCart(p0[0] as HashMap<String, String>)
+        }
+
+        override fun onPostExecute(result: JSONObject?) {
+            if (result!!.optInt("code", 0) != 0) {
+
+                //context.longToast(result.toString())
+                context.longToast("${productName} Modified From Cart")
+
+                EventBus.getDefault().post(ItemRemovedFromCartEvent(true))
+            }
+            super.onPostExecute(result)
+        }
     }
 
     fun deleteFromCart() {
+
         if (context.isConnected()) {
-            AsyncTaskCompat.executeParallel(DeleteFromCart(), null)
+
+            val builder = AlertDialog.Builder(context)
+            builder.setTitle("Are You Sure?")
+            builder.setMessage("Are you sure you want to remove ${productName} from cart ...")
+            builder.setPositiveButton("OK", { dialogInterface, i ->
+                run {
+                    AsyncTaskCompat.executeParallel(DeleteFromCart(), null)
+                }
+            })
+            builder.setNegativeButton("Cancel", null)
+            builder.show()
+
+        } else {
+            context.toast("No Internet Connection")
         }
     }
 
@@ -62,7 +151,7 @@ class ShoppingCartProductAdapter(
         override fun onPostExecute(result: JSONObject?) {
 
             if (result!!.optInt("code", 0) != 0) {
-                context.toast("Item Removed From Cart")
+                context.toast("${productName} Removed From Cart")
 
                 EventBus.getDefault().post(ItemRemovedFromCartEvent(true))
             }
